Bug 28428: Capitalization: Password Updated
This patch corrects the case of two headings on the "Change password"
page in the staff interface.

As far as I can tell these strings are never seen because the page
redirects to the patron detail page instead of redisplaying after a new
password or username is submitted.

A visual inspection of the patch may be the only test.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>