Bug 28175: (follow-up) Preselect correct category after search
This patch modifies the script and template so that if the user submits
a search by category, that category will be preselected on the search
results page.

Also added is an empty "Choose" option so that a search category isn't
selected by default.

To test, apply the patch and view the uploads page.

 - In the category search form, a "Choose" option should be selected.
   Submitting the form without selecting a category should not work.
 - Search using one of your existing categories. On the search results
   page the category you selected should remain selected.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>